+++ include/pbcopper/cli2/internal/BuiltinOptions.h
@@ -15,7 +15,7 @@ namespace Builtin {
 const Option Help{
 R"({
     "names" : ["h", "help"],
-    "description" : "Show this help."
+    "description" : "Show this help and exit."
 })"};
 
 const Option LogLevel{
@@ -24,9 +24,8 @@ R"({
     "names.hidden" : ["logLevel"],
     "description" : "Set log level.",
     "type" : "string",
-    "default" : "INFO",
-    "choices" : ["TRACE", "DEBUG", "INFO", "WARN", "FATAL"],
-    "choices.hidden" : true
+    "default" : "WARN",
+    "choices" : ["TRACE", "DEBUG", "INFO", "WARN", "FATAL"]
 })"};
 
 const Option LogFile{
@@ -34,13 +33,22 @@ R"({
     "names" : ["log-file"],
     "names.hidden" : ["logFile"],
     "description" : "Log to a file, instead of stderr.",
-    "type" : "file"
+    "type" : "file",
+    "default" : ""
+})"};
+
+const Option NumThreads{
+R"({
+    "names" : ["j", "num-threads"],
+    "description" : "Number of threads to use, 0 means autodetection.",
+    "type" : "unsigned int",
+    "default" : 0
 })"};
 
 const Option Version{
 R"({
     "names" : ["version"],
-    "description" : "Show application version."
+    "description" : "Show application version and exit."
 })"};
 
 } // namespace Builtin