+++ src/cli2/Results.cpp
@@ -1,7 +1,10 @@
 #include <pbcopper/cli2/Results.h>
 
+#include <algorithm>
 #include <stdexcept>
+#include <thread>
 
+#include <pbcopper/cli2/internal/BuiltinOptions.h>
 #include <pbcopper/cli2/internal/OptionTranslator.h>
 #include <pbcopper/cli2/internal/PositionalArgumentTranslator.h>
 
@@ -55,20 +58,30 @@ Results& Results::InputCommandLine(std::string cmdLine)
     return *this;
 }
 
-PacBio::Logging::LogLevel Results::LogLevel() const { return logLevel_; }
-
-Results& Results::LogLevel(PacBio::Logging::LogLevel logLevel)
+std::string Results::LogFile() const
 {
-    logLevel_ = logLevel;
-    return *this;
+    const auto& logFileOpt = (*this)[Builtin::LogFile];
+    const std::string logFileStr = logFileOpt;
+    return logFileStr;
 }
 
-size_t Results::NumProcessors() const { return numProcessors_; }
-
-Results& Results::NumProcessors(size_t nproc)
+PacBio::Logging::LogLevel Results::LogLevel() const
 {
-    numProcessors_ = nproc;
-    return *this;
+    const auto& logLevelOpt = (*this)[Builtin::LogLevel];
+    const std::string logLevelStr = logLevelOpt;
+    return PacBio::Logging::LogLevel(logLevelStr);
+}
+
+size_t Results::NumThreads() const
+{
+    const auto& numThreadOpt = (*this)[Builtin::NumThreads];
+    const unsigned int requestedNumThreads = numThreadOpt;
+    // NOTE: max may be 0 if unknown
+    const unsigned int maxNumThreads = std::thread::hardware_concurrency();
+
+    if (requestedNumThreads == 0) return std::max(1U, maxNumThreads);
+    if (maxNumThreads == 0) return requestedNumThreads;
+    return std::min(requestedNumThreads, maxNumThreads);
 }
 
 const std::vector<std::string>& Results::PositionalArguments() const { return posArgValues_; }